+++ src/design/plone/contenttypes/browser/utils/check_luoghi.py
@@ -1,3 +1,4 @@
+from DateTime import DateTime
 from openpyxl import Workbook
 from openpyxl.styles import Alignment
 from openpyxl.styles import Font
@@ -12,23 +13,30 @@ import io
 FLAG = '<i class="fa-solid fa-check"></i>'
 
 
-class CheckNotizie(BrowserView):
+class CheckLuoghi(BrowserView):
     cds = None
 
     def is_anonymous(self):
         return api.user.is_anonymous()
 
-    def information_dict(self, notizia):
+    def information_dict(self, luogo):
+        indirizzo = False
+        if getattr(luogo, "city", "") and luogo.city.strip():
+            if getattr(luogo, "street", "") and luogo.street.strip():
+                if getattr(luogo, "zip_code", "") and luogo.zip_code.strip():
+                    indirizzo = True
 
-        descrizione_estesa = getattr(notizia, "descrizione_estesa", "")
-        res = [x.get("text", "") for x in descrizione_estesa["blocks"].values()]
+        modalita_accesso = getattr(luogo, "modalita_accesso", "")
+        res = [x.get("text", "") for x in modalita_accesso["blocks"].values()]
         if not [x for x in res if x]:
-            descrizione_estesa = ""
+            modalita_accesso = ""
 
         return {
-            "descrizione_estesa": descrizione_estesa,
-            "effective_date": getattr(notizia, "effective_date", None),
-            "a_cura_di": getattr(notizia, "a_cura_di", None),
+            "description": getattr(luogo, "description", "").strip(),
+            "image": getattr(luogo, "image", None),
+            "indirizzo": indirizzo,
+            "modalita_accesso": modalita_accesso,
+            "contact_info": getattr(luogo, "contact_info", None),
         }
 
     def plone2volto(self, url):
@@ -40,45 +48,48 @@ class CheckNotizie(BrowserView):
             return url.replace(portal_url, frontend_domain, 1)
         return url
 
-    def get_notizie(self):
+    def get_luoghi(self):
+
         if self.is_anonymous():
             return []
         pc = api.portal.get_tool("portal_catalog")
 
+        # show_inactive ha sempre avuto una gestione... particolare! aggiungo ai
+        # kw effectiveRange = DateTime() che è quello che fa Products.CMFPlone
+        # nel CatalogTool.py
         query = {
-            "portal_type": "News Item",
+            "portal_type": "Venue",
             "review_state": "published",
         }
-
-        brains = pc(query)
+        brains = pc(query, **{"effectiveRange": DateTime()})
         results = {}
         for brain in brains:
-            notizia = brain.getObject()
+            luogo = brain.getObject()
 
-            information_dict = self.information_dict(notizia)
-
+            information_dict = self.information_dict(luogo)
             if all(information_dict.values()):
                 continue
 
-            parent = notizia.aq_inner.aq_parent
+            parent = luogo.aq_inner.aq_parent
             if parent.title not in results:
                 results[parent.title] = {
                     "url": self.plone2volto(parent.absolute_url()),
                     "children": [],
                 }
-
             results[parent.title]["children"].append(
                 {
-                    "title": notizia.title,
-                    "descrizione_estesa": information_dict.get("descrizione_estesa")
-                    and FLAG
-                    or "",
-                    "url": self.plone2volto(notizia.absolute_url()),
+                    "title": luogo.title,
+                    "description": information_dict.get("description") and FLAG or "",
+                    "url": self.plone2volto(luogo.absolute_url()),
                     "data": {
-                        "effective_date": information_dict.get("effective_date")
+                        "image": information_dict.get("image") and FLAG or "",
+                        "indirizzo": information_dict.get("indirizzo") and FLAG or "",
+                        "modalita_accesso": information_dict.get("modalita_accesso")
                         and FLAG
                         or "",
-                        "a_cura_di": information_dict.get("a_cura_di") and FLAG or "",
+                        "contact_info": information_dict.get("contact_info")
+                        and FLAG
+                        or "",
                     },
                 }
             )
@@ -90,20 +101,22 @@ class CheckNotizie(BrowserView):
         return results
 
 
-class DownloadCheckNotizie(CheckNotizie):
+class DownloadCheckLuoghi(CheckLuoghi):
     CT = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
 
     def __call__(self):
         HEADER = [
             "Titolo",
-            "Descrizione estesa",
-            "Data di pubblicazione",
-            "A cura di",
+            "Descrizione",
+            "Immagine in evidenza",
+            "Indirizzo",
+            "Modalità do accesso",
+            "Contatti",
         ]
 
         EMPTY_ROW = [""] * 3
 
-        notizie = self.get_notizie()
+        luoghi = self.get_luoghi()
 
         workbook = Workbook()
         sheet = workbook.active
@@ -117,7 +130,7 @@ class DownloadCheckNotizie(CheckNotizie):
 
         section_row_height = int(14 * 1.5)
 
-        for category, category_data in notizie.items():
+        for category, category_data in luoghi.items():
             section_url = category_data["url"]
             section_title = category
             section_row = [section_title, "", ""]
@@ -150,15 +163,17 @@ class DownloadCheckNotizie(CheckNotizie):
                 column_letter = get_column_letter(col[0].column)
                 sheet.column_dimensions[column_letter].width = 35
 
-            for notizia in category_data["children"]:
-                title_url = notizia["url"]
-                dati_notizia = [
-                    notizia["title"],
-                    "X" if notizia["descrizione_estesa"] else "",
-                    "X" if notizia["data"]["effective_date"] else "",
-                    "X" if notizia["data"]["a_cura_di"] else "",
+            for luogo in category_data["children"]:
+                title_url = luogo["url"]
+                dati_luogo = [
+                    luogo["title"],
+                    "X" if luogo["description"] else "",
+                    "X" if luogo["data"]["image"] else "",
+                    "X" if luogo["data"]["indirizzo"] else "",
+                    "X" if luogo["data"]["modalita_accesso"] else "",
+                    "X" if luogo["data"]["contact_info"] else "",
                 ]
-                row = dati_notizia
+                row = dati_luogo
                 sheet.append(row)
 
                 title_cell = sheet.cell(row=sheet.max_row, column=1)
@@ -179,6 +194,6 @@ class DownloadCheckNotizie(CheckNotizie):
         self.request.RESPONSE.setHeader("Content-Type", self.CT)
         self.request.response.setHeader(
             "Content-Disposition",
-            "attachment; filename=check_notizie.xlsx",
+            "attachment; filename=check_luoghi.xlsx",
         )
         return data