+++ src/design/plone/contenttypes/browser/utils/check_uo.py
@@ -1,3 +1,4 @@
+from DateTime import DateTime
 from openpyxl import Workbook
 from openpyxl.styles import Alignment
 from openpyxl.styles import Font
@@ -5,6 +6,9 @@ from openpyxl.styles import PatternFill
 from openpyxl.utils import get_column_letter
 from plone import api
 from Products.Five import BrowserView
+from zope.globalrequest import getRequest
+from zope.component import getMultiAdapter
+from plone.restapi.interfaces import ISerializeToJsonSummary
 
 import io
 
@@ -12,23 +16,44 @@ import io
 FLAG = '<i class="fa-solid fa-check"></i>'
 
 
-class CheckNotizie(BrowserView):
+class CheckUnitaOrganizzative(BrowserView):
     cds = None
 
     def is_anonymous(self):
         return api.user.is_anonymous()
 
-    def information_dict(self, notizia):
+    def get_relations(self, obj, field):
+        return api.relation.get(source=obj, relationship=field, unrestricted=False)
 
-        descrizione_estesa = getattr(notizia, "descrizione_estesa", "")
-        res = [x.get("text", "") for x in descrizione_estesa["blocks"].values()]
+    def get_related_objects(self, obj, field):
+        """ """
+        items = []
+        relations = self.get_relations(obj, field)
+
+        for rel in relations:
+            rel_obj = rel.to_object
+            if rel_obj is not None:
+                summary = getMultiAdapter(
+                    (rel_obj, getRequest()), ISerializeToJsonSummary
+                )()
+                items.append(summary)
+        return sorted(items, key=lambda k: k["title"])
+
+    def information_dict(self, uo):
+        sede_ref = None
+        sede_ref = self.get_related_objects(uo, "sede")
+        if sede_ref:
+            sede_ref = sede_ref[0]
+
+        competenze = getattr(uo, "competenze", "")
+        res = [x.get("text", "") for x in competenze["blocks"].values()]
         if not [x for x in res if x]:
-            descrizione_estesa = ""
-
+            competenze = ""
         return {
-            "descrizione_estesa": descrizione_estesa,
-            "effective_date": getattr(notizia, "effective_date", None),
-            "a_cura_di": getattr(notizia, "a_cura_di", None),
+            "description": getattr(uo, "description", "").strip(),
+            "competenze": competenze,
+            "sede": sede_ref,
+            "contact_info": getattr(uo, "contact_info", None),
         }
 
     def plone2volto(self, url):
@@ -40,45 +65,46 @@ class CheckNotizie(BrowserView):
             return url.replace(portal_url, frontend_domain, 1)
         return url
 
-    def get_notizie(self):
+    def get_uos(self):
+
         if self.is_anonymous():
             return []
         pc = api.portal.get_tool("portal_catalog")
 
+        # show_inactive ha sempre avuto una gestione... particolare! aggiungo ai
+        # kw effectiveRange = DateTime() che è quello che fa Products.CMFPlone
+        # nel CatalogTool.py
         query = {
-            "portal_type": "News Item",
+            "portal_type": "UnitaOrganizzativa",
             "review_state": "published",
         }
-
-        brains = pc(query)
+        brains = pc(query, **{"effectiveRange": DateTime()})
         results = {}
         for brain in brains:
-            notizia = brain.getObject()
+            uo = brain.getObject()
 
-            information_dict = self.information_dict(notizia)
+            information_dict = self.information_dict(uo)
 
             if all(information_dict.values()):
                 continue
 
-            parent = notizia.aq_inner.aq_parent
+            parent = uo.aq_inner.aq_parent
             if parent.title not in results:
                 results[parent.title] = {
                     "url": self.plone2volto(parent.absolute_url()),
                     "children": [],
                 }
-
             results[parent.title]["children"].append(
                 {
-                    "title": notizia.title,
-                    "descrizione_estesa": information_dict.get("descrizione_estesa")
-                    and FLAG
-                    or "",
-                    "url": self.plone2volto(notizia.absolute_url()),
+                    "title": uo.title,
+                    "description": information_dict.get("description") and FLAG or "",
+                    "url": self.plone2volto(uo.absolute_url()),
                     "data": {
-                        "effective_date": information_dict.get("effective_date")
+                        "competenze": information_dict.get("competenze") and FLAG or "",
+                        "sede": information_dict.get("sede") and FLAG or "",
+                        "contact_info": information_dict.get("contact_info")
                         and FLAG
                         or "",
-                        "a_cura_di": information_dict.get("a_cura_di") and FLAG or "",
                     },
                 }
             )
@@ -90,20 +116,15 @@ class CheckNotizie(BrowserView):
         return results
 
 
-class DownloadCheckNotizie(CheckNotizie):
+class DownloadCheckUnitaOrganizzative(CheckUnitaOrganizzative):
     CT = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
 
     def __call__(self):
-        HEADER = [
-            "Titolo",
-            "Descrizione estesa",
-            "Data di pubblicazione",
-            "A cura di",
-        ]
+        HEADER = ["Titolo", "Descrizione", "Competenze", "Sede", "Contatti"]
 
         EMPTY_ROW = [""] * 3
 
-        notizie = self.get_notizie()
+        unita_organizzative = self.get_uos()
 
         workbook = Workbook()
         sheet = workbook.active
@@ -117,7 +138,7 @@ class DownloadCheckNotizie(CheckNotizie):
 
         section_row_height = int(14 * 1.5)
 
-        for category, category_data in notizie.items():
+        for category, category_data in unita_organizzative.items():
             section_url = category_data["url"]
             section_title = category
             section_row = [section_title, "", ""]
@@ -150,15 +171,16 @@ class DownloadCheckNotizie(CheckNotizie):
                 column_letter = get_column_letter(col[0].column)
                 sheet.column_dimensions[column_letter].width = 35
 
-            for notizia in category_data["children"]:
-                title_url = notizia["url"]
-                dati_notizia = [
-                    notizia["title"],
-                    "X" if notizia["descrizione_estesa"] else "",
-                    "X" if notizia["data"]["effective_date"] else "",
-                    "X" if notizia["data"]["a_cura_di"] else "",
+            for uo in category_data["children"]:
+                title_url = uo["url"]
+                dati_uo = [
+                    uo["title"],
+                    "X" if uo["description"] else "",
+                    "X" if uo["data"]["competenze"] else "",
+                    "X" if uo["data"]["sede"] else "",
+                    "X" if uo["data"]["contact_info"] else "",
                 ]
-                row = dati_notizia
+                row = dati_uo
                 sheet.append(row)
 
                 title_cell = sheet.cell(row=sheet.max_row, column=1)
@@ -179,6 +201,6 @@ class DownloadCheckNotizie(CheckNotizie):
         self.request.RESPONSE.setHeader("Content-Type", self.CT)
         self.request.response.setHeader(
             "Content-Disposition",
-            "attachment; filename=check_notizie.xlsx",
+            "attachment; filename=check_unita_organizzative.xlsx",
         )
         return data